Update popup.js
ADD extension disable refresh page

If the extension is turned off, the page will also be refreshed. Adding a judgment condition will only refresh the Berriz page. If the number of pages is greater than 1, use the queue to refresh all the Berriz.in pages that meet the conditions in sequence.